app/requirements.txt
python-iptables==0.14.0
PyYAML==5.3
